Add files via upload
The below content describes the process of creating collections and querying data using Apache Solr for two datasets: "Films" and "Books." Solr is used as a powerful search platform for indexing and querying large datasets. The steps involved in creating collections, editing schema, loading data, and executing queries are explained for both datasets.